+++ test/Rule/PhoneRuleTest.php
@@ -8,6 +8,7 @@ use DateTimeImmutable;
 use Monolog\Level;
 use Monolog\LogRecord;
 use PHPUnit\Framework\TestCase;
+use Sirix\Monolog\Redaction\Exception\RedactorReflectionException;
 use Sirix\Monolog\Redaction\RedactorProcessor;
 use Sirix\Monolog\Redaction\Rule\PhoneRule;
 use Test\Sirix\Monolog\Redaction\NestedArrayConversionTrait;
@@ -16,6 +17,9 @@ final class PhoneRuleTest extends TestCase
 {
     use NestedArrayConversionTrait;
 
+    /**
+     * @throws RedactorReflectionException
+     */
     public function testPhoneNumberRedaction(): void
     {
         $rule = new PhoneRule();
@@ -26,6 +30,9 @@ final class PhoneRuleTest extends TestCase
         $this->assertSame('1234****90', $processed->context['phone']);
     }
 
+    /**
+     * @throws RedactorReflectionException
+     */
     public function testInternationalPhoneNumberRedaction(): void
     {
         $rule = new PhoneRule();
@@ -36,6 +43,9 @@ final class PhoneRuleTest extends TestCase
         $this->assertSame('+4412****12', $processed->context['phone']);
     }
 
+    /**
+     * @throws RedactorReflectionException
+     */
     public function testFormattedPhoneNumberRedaction(): void
     {
         $rule = new PhoneRule();
@@ -46,6 +56,9 @@ final class PhoneRuleTest extends TestCase
         $this->assertSame('123-******90', $processed->context['phone']);
     }
 
+    /**
+     * @throws RedactorReflectionException
+     */
     public function testPhoneRedactionInNestedStructures(): void
     {
         $rule = new PhoneRule();
@@ -70,6 +83,9 @@ final class PhoneRuleTest extends TestCase
         $this->assertSame('9876****10', $processed->context['user']->contact->phone);
     }
 
+    /**
+     * @throws RedactorReflectionException
+     */
     public function testNonPhoneValue(): void
     {
         $rule = new PhoneRule();
@@ -80,6 +96,9 @@ final class PhoneRuleTest extends TestCase
         $this->assertSame('This********************er', $processed->context['value']);
     }
 
+    /**
+     * @throws RedactorReflectionException
+     */
     public function testShortPhoneNumber(): void
     {
         $rule = new PhoneRule();